requirements.txt
galaxy-util==25.0.4
lxml==6.0.2
